requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
requests==2.32.2
python-dotenv==1.0.1
urllib3==2.2.1
gunicorn==23.0.0
Werkzeug==3.0.3
blinker==1.8.2
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7